+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to **RTVI Client Web** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2024-08-28
+
+### Fixed
+
+- All config mutation methods (getServiceOptionsFromConfig, getServiceOptionValueFromConfig, setServiceOptionInConfig) now work when not in a ready state.
+
+### Added
+
+- New config method: `getServiceOptionValueFromConfig`. Returns value of config service option with passed service key and option name.
+- setServiceOptionInConfig now accepts either one or many ConfigOption arguments (and will set or update all)
+- setServiceOptionInConfig now accepts an optional 'config' param, which it will use over the default VoiceClient config. Useful if you want to mutate an existing config option across multiple services before calling `updateConfig`.
+- New config method `setConfigOptions` updates multiple service options by running each item through `setServiceOptionInConfig`.
+
+### Fixed
+
+- "@daily-co/daily-js" should not be included in the `rtvi-client-js` package.json. This dependency is only necessary for `rtvi-client-js-daily`.
+- Jest unit tests added for config manipulation within `rtvi-client-js` (`yarn run test`)
+
+## [0.1.6] - 2024-08-26
+
+### Fixed
+
+- `getServiceOptionsFromConfig` should return a new object, not an instance of the config. This prevents methods like `setContext` from mutating local config unintentionally.
+
 ## [0.1.5] - 2024-08-19
 
 ### Added